refactor: [UNM/API-25] file based routing with full route separation
